+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Validation for allowing extensionless attachments
+
 ### Fixed
 
 - Year being returned from Intl date formatter as only digits in iOS15